requirements.txt
bootstrap-flask
flask
flask-login
flask-sqlalchemy
flask-wtf
pillow
gunicorn
pycrypto
